add weighing unitest
